Postgres-AgenticAI-LangGraph/requirements.txt
psycopg2
psycopg2-binary
pandas
python-dotenv
langchain_community
pyprojroot
openai
google-generativeai
langchain_google_genai
langchain_openai
gradio